moved campus logic to utils
