+++ modules/calendar/calendar_creator.py
@@ -1,8 +1,7 @@
-from typing import Iterable
 from utils.sql_fetcher import SqlFetcher
 import psycopg2.extensions as sql
 from .calendar_service import CalendarService
-from .campus import Campus
+from utils.campus.campus import Campus
 
 
 class CalendarCreator:
@@ -13,24 +12,17 @@ class CalendarCreator:
 		self.__sql_fetcher = sql_fetcher
 		self.__service = service
 
-	async def create_class_calendars(self, year: int):
+	def create_class_calendars(self, year: int):
 		"""Create a calendar for each campus and add it to the database"""
-		for campus in await self.__get_campuses():
+		for campus in Campus.get_campuses():
 			# create calendar
-			calendar = self.__service.create_calendar(f"JCT CompSci {campus.name} {year}")
+			calendar = self.__service.create_calendar(
+				f"JCT CompSci {campus.name} {year}"
+			)
 			# update class in database
-			await self.__add_calendar(calendar.id, campus.id, year)
+			self.__add_calendar(calendar.id, campus.id, year)
 
-	async def __get_campuses(self) -> Iterable[Campus]:
-		"""Fetch a list of campuses from the database"""
-		query = self.__sql_fetcher["get_campuses.sql"]
-		with self.__conn as conn:
-			with conn.cursor() as cursor:
-				cursor.execute(query)
-				campuses = [Campus(*tup) for tup in cursor.fetchall()]
-		return campuses
-
-	async def __add_calendar(self, calendar_id: str, campus_id: int, year: int):
+	def __add_calendar(self, calendar_id: str, campus_id: int, year: int):
 		"""Update the class entry in the database with the calendar id"""
 		query = self.__sql_fetcher["add_calendar.sql"]
 		with self.__conn as conn:
