--- modules/create_class/campus.py
@@ -1,4 +1,0 @@
-class Campus:
-	def __init__(self, id: int, name: str):
-		self.id = id
-		self.name = name
