--- modules/create_class/queries/get_campuses.sql
@@ -1,1 +1,0 @@
-SELECT id, name FROM campuses
