+++ utils/campus/queries/get_campuses.sql
